chore: move StateIdentifiers to Domian layer
