+++ src/Application/types.ts
@@ -6,12 +6,6 @@ export type Identifier =  {
 
 export type IdentifierBuilder = (scope: IdentifierScope) => Identifier;
 
-export type StateIdentifiers = {
-    apikey: string;
-    ip?: string;
-    tenant?: string;
-};
-
 export type UserIdentity = {
     apiKey: string;
     ip?: string;
